This PR picks the fastest (up to now) version of filtering: group conv for small filters, reshape for large ones.

There are some cases, when (for 9x9 or 11x11 and depending on hardware) kernel, the selection is not optimal, but it is the best one I got with a simple condition.

Benchmark similar to https://github.com/kornia/kornia/issues/355 but with backwards:

```python3
import torch
import random
import torch.nn.functional as F
import torch.backends.cudnn as cudnn
import timeit
def benchmark_group(input, device, sh):
    torch.cuda.synchronize()
    x = F.conv2d(input, kernel.expand(c,-1,-1,-1), groups=c)
    loss = x.sum()
    loss.backward()
    torch.cuda.synchronize()
    return

def benchmark_reshape(input, device, sh):
    b1,c1,h1,w1 = input.size()
    torch.cuda.synchronize()
    x = F.conv2d(input.view(b*c,1, h1, w1), kernel).view(*sh)
    loss = x.sum()
    loss.backward()
    torch.cuda.synchronize()
    return


def benchmark_adapt(input, device, sh):
    b1,c1,h1,w1 = input.size()
    torch.cuda.synchronize()
    dev = input.device
    ks_numel = kernel.size(3) * kernel.size(2)
    if (ks_numel > 81):
        x = F.conv2d(input.view(b*c,1, h1,w1), kernel).view(*sh)
    else:
        x =  F.conv2d(input, kernel.expand(c,-1,-1,-1), groups=c)        
    loss = x.sum()
    loss.backward()
    torch.cuda.synchronize()
    return

#For square kernels
b,c,h,w = 16,16,800,800
sh = [b,c,h,w]
for ks in [3, 5,7, 9, 11, 13, 15]: 
    for device in [torch.device('cuda:0')]:# torch.device('cpu')]: 
        data = torch.rand(b,c,h,w, device=device).float() 
        x = F.pad(data,(ks//2, ks//2, ks//2, ks//2)) 
        b1,c1,h1,w1 = x.size() 
        kernel = torch.rand(1,1,ks,ks, device=device, requires_grad=True).float() 
        #warmup 
        o = F.conv2d(x, kernel.expand(c,-1,-1,-1), groups=c)
        o = F.conv2d(x.view(b*c,1, h1,w1), kernel).view(b,c,h,w) 
        torch.cuda.synchronize() 
        print (ks, 'group', str(device)) 
        %timeit benchmark_group(x, device, sh) 
        kernel = torch.rand(1,1,ks,ks, device=device, requires_grad=True).float() 
        torch.cuda.synchronize() 
        print (ks, 'reshape',str(device))
        kernel = torch.rand(1,1,ks,ks, device=device, requires_grad=True).float() 
        %timeit benchmark_reshape(x,device, sh)
        kernel = torch.rand(1,1,ks,ks, device=device, requires_grad=True).float() 
        print (ks, 'adapt',str(device))
        %timeit benchmark_adapt(x,device, sh)
    print ("   ")

#For 1xn kernels

b,c,h,w = 16,16,800,800
sh = [b,c,h,w]

for ks in [3, 5,7, 9, 11, 13, 15, 49]: 
    for device in [torch.device('cuda:0')]:# torch.device('cpu')]: 
        data = torch.rand(b,c,h,w, device=device).float() 
        x = F.pad(data,(ks//2,ks//2,0,0)) 
        print (x.shape)
        b1,c1,h1,w1 = x.size() 
        kernel = torch.rand(1,1, 1, ks, device=device, requires_grad=True).float() 
        #warmup 
        o = F.conv2d(x, kernel.expand(c,-1,-1,-1), groups=c)
        o = F.conv2d(x.view(b*c,1, h1, w1), kernel).view(b,c,h,w) 
        torch.cuda.synchronize() 
        print (ks, 'group', str(device)) 
        %timeit benchmark_group(x,device, sh) 
        kernel = torch.rand(1,1, 1, ks, device=device, requires_grad=True).float() 
        torch.cuda.synchronize() 
        print (ks, 'reshape',str(device))
        kernel = torch.rand(1,1, 1, ks, device=device, requires_grad=True).float() 
        %timeit benchmark_reshape(x,device,sh)
        kernel = torch.rand(1,1, 1, ks, device=device, requires_grad=True).float() 
        print (ks, 'adapt',str(device))
        %timeit benchmark_adapt(x,device,sh)
    print ("   ")
```

Results: square

> 3 group cuda:0
> 49.5 ms ± 167 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 3 reshape cuda:0
> 193 ms ± 483 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 3 adapt cuda:0
> 50 ms ± 143 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 5 group cuda:0
> 168 ms ± 554 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 5 reshape cuda:0
> 215 ms ± 4.64 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 5 adapt cuda:0
> 167 ms ± 397 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 7 group cuda:0
> 297 ms ± 4.26 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 7 reshape cuda:0
> 241 ms ± 1.17 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 7 adapt cuda:0
> 304 ms ± 3.02 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>    
> 9 group cuda:0
> 455 ms ± 7.82 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 9 reshape cuda:0
> 467 ms ± 3.61 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 9 adapt cuda:0
> 449 ms ± 5.15 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>    
> 11 group cuda:0
> 618 ms ± 9.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 11 reshape cuda:0
> 508 ms ± 1.65 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 11 adapt cuda:0
> 508 ms ± 1.44 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>    
> 13 group cuda:0
> 914 ms ± 8.26 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 13 reshape cuda:0
> 720 ms ± 1.66 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 13 adapt cuda:0
> 723 ms ± 1.34 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>    
> 15 group cuda:0
> 1.28 s ± 18.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 15 reshape cuda:0
> 946 ms ± 1.54 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 15 adapt cuda:0
> 948 ms ± 1.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Rectangular:


> 3 group cuda:0
> 29.6 ms ± 4.85 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 3 reshape cuda:0
> 183 ms ± 1.16 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 3 adapt cuda:0
> 29.7 ms ± 2.91 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 
> 5 group cuda:0
> 36.5 ms ± 6.51 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 5 reshape cuda:0
> 190 ms ± 875 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 5 adapt cuda:0
> 36.5 ms ± 4.17 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 
> 7 group cuda:0
> 43.9 ms ± 10.4 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 7 reshape cuda:0
> 190 ms ± 929 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 7 adapt cuda:0
> 43.9 ms ± 5.46 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 
> 9 group cuda:0
> 51 ms ± 48.6 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 9 reshape cuda:0
> 194 ms ± 876 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 9 adapt cuda:0
> 50.9 ms ± 118 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 
> 11 group cuda:0
> 60.5 ms ± 624 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 11 reshape cuda:0
> 195 ms ± 1.14 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 11 adapt cuda:0
> 60.9 ms ± 490 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 
> 13 group cuda:0
> 96.4 ms ± 623 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 13 reshape cuda:0
> 198 ms ± 766 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 13 adapt cuda:0
> 96.8 ms ± 533 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 15 group cuda:0
> 118 ms ± 420 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 15 reshape cuda:0
> 198 ms ± 395 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> 15 adapt cuda:0
> 117 ms ± 977 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
>    
> 49 group cuda:0
> 294 ms ± 11.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 49 reshape cuda:0
> 237 ms ± 2.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> 49 adapt cuda:0
> 291 ms ± 7.25 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>