@ducha-aiki 2020-01-29
> LGTM. Can we integrate the code you implemented for bench-marking as performance tests ?

Sure. Where do you think is the best place for performance test?