@edgarriba 2020-01-29
somewhere under tests dir. Maybe we could use pytest mark to explicitli run them